AI_Builders/requirement.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
idna==3.2
Pillow==8.3.1
pkg_resources==0.0.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
psycopg2==2.8.6
